:recycle: Make summary panel read only, move in territory selector into purpose panel.
