--- web/src/modules/travel-authorizations/pages/travel-authorization-edit/SummaryHeaderForm.vue
@@ -1,166 +1,0 @@
-<template>
-  <v-form
-    ref="form"
-    lazy-validation
-  >
-    <v-row dense>
-      <v-col
-        cols="12"
-        md="2"
-      >
-        <!-- Depending on in territory flag we will load a different list of destinations -->
-        <v-checkbox
-          v-model="currentTravelAuthorization.allTravelWithinTerritory"
-          label="In Territory?"
-          dense
-          required
-        >
-        </v-checkbox>
-      </v-col>
-    </v-row>
-    <v-row dense>
-      <v-col
-        cols="12"
-        md="2"
-      >
-        <v-select
-          v-model="currentTravelAuthorization.purposeId"
-          :items="purposes"
-          :loading="loadingPurposes"
-          :rules="[required]"
-          background-color="white"
-          dense
-          item-text="purpose"
-          item-value="id"
-          label="Purpose"
-          outlined
-          required
-        ></v-select>
-      </v-col>
-      <v-col
-        cols="12"
-        md="2"
-      >
-        <v-autocomplete
-          v-model="finalDestination.locationId"
-          :items="destinationsByCurrentFormTravelRestriction"
-          :loading="loadingDestinations"
-          :rules="[required]"
-          background-color="white"
-          clearable
-          dense
-          label="Final Destination"
-          outlined
-          persistent-hint
-          required
-          validate-on-blur
-        >
-        </v-autocomplete>
-      </v-col>
-      <v-col
-        cols="12"
-        md="2"
-      >
-        <DatePicker
-          v-model="initialDestination.departureDate"
-          :rules="[required]"
-          label="Start Date"
-          required
-        />
-      </v-col>
-      <v-col
-        cols="12"
-        md="2"
-      >
-        <DatePicker
-          v-model="finalDestination.departureDate"
-          :min="initialDestination.departureDate"
-          :rules="[
-            required,
-            greaterThanOrEqualToDate(initialDestination.departureDate, {
-              referenceFieldLabel: 'start date',
-            }),
-          ]"
-          label="End Date"
-          required
-        />
-      </v-col>
-    </v-row>
-  </v-form>
-</template>
-
-<script>
-import { first, last } from "lodash"
-import { mapState, mapActions, mapGetters } from "vuex"
-
-import { required, greaterThanOrEqualToDate } from "@/utils/validators"
-
-import DatePicker from "@/components/Utils/DatePicker"
-
-export default {
-  name: "SummaryHeaderForm",
-  components: {
-    DatePicker,
-  },
-  data: () => ({
-    loadingPurposes: false,
-    loadingDestinations: false,
-  }),
-  computed: {
-    ...mapState("travelAuthorizations", [
-      "currentTravelAuthorization",
-      "purposes",
-      "destinationsByCurrentFormTravelRestriction",
-    ]),
-    ...mapGetters("travelAuthorizations", [
-      "currentTravelAuthorizationId",
-      "destinationsByCurrentFormTravelRestriction",
-    ]),
-    finalDestination: {
-      get() {
-        return (
-          last(this.currentTravelAuthorization.stops) || {
-            travelAuthorizationId: this.currentTravelAuthorizationId,
-          }
-        )
-      },
-      set(newValue) {
-        this.$set(
-          this.currentTravelAuthorization.stops,
-          this.currentTravelAuthorization.stops.length - 1,
-          newValue
-        )
-      },
-    },
-    initialDestination: {
-      get() {
-        return (
-          first(this.currentTravelAuthorization.stops) || {
-            travelAuthorizationId: this.currentTravelAuthorizationId,
-          }
-        )
-      },
-      set(newValue) {
-        this.$set(this.currentTravelAuthorization.stops, 0, newValue)
-      },
-    },
-  },
-  async mounted() {
-    this.loadingPurposes = true
-    this.loadingDestinations = true
-    await Promise.all([
-      this.loadPurposes().finally(() => {
-        this.loadingPurposes = false
-      }),
-      this.loadDestinations().finally(() => {
-        this.loadingDestinations = false
-      }),
-    ])
-  },
-  methods: {
-    ...mapActions("travelAuthorizations", ["loadPurposes", "loadDestinations"]),
-    required,
-    greaterThanOrEqualToDate,
-  },
-}
-</script>
